Validate stream has data
Created the "UploadedStreamValidator" class to handle validating uploaded file streams. Currently it only has an "EnsureStreamHasData" method that is intended to ensure the stream has data in it.

Currently the new method will always throw an exception since it is a work in progress.
Issue: https://github.com/TownSuite/TownSuite.IssueTracker/issues/7288