Don't validate stream has data when seek unavailable
An exception will no longer be thrown by the "EnsureStreamHasData" method when seek is unavailable.

In production, the length to check would of been for a NetworkStream which doesn't support Seek/Length. This seems like it would make the length check pointless, however this class is still a work in progress. To make this a viable option the UploadStreamValidator class was renamed to UploadStreamHandler and is in charge of more responsiblities. It may be broken down into more classes later. The class will do more than validate the stream later on, it will also copy from 1 stream to another.
Issue: https://github.com/TownSuite/TownSuite.IssueTracker/issues/7288