Stream copying validates length of result
The "CopyToAsync" method will now handle the validation of the stream length. It will only validate the resulting stream to give a better exception message. This will not validate the from stream since in production a NetworkStream is expected that doesn't support length.

The "EnsureStreamHasData" method was made private and the tests for it now use "CopyToAsync" instead. The test for seek is no longer needed since it is unexpected for the stream being copied to to not support seek. An exception would of been thrown before EnsureStreamHasData is even called due to an earlier Seek.
Issue: https://github.com/TownSuite/TownSuite.IssueTracker/issues/7288