Ensure stream cannot be too large
A maximum stream size can be given to the CopyToAsync method of the UploadedStreamHandler class. The method will throw an exception if the Stream to copy from has more byte's after copying the maximum amount of bytes.

The CopyToAsync method can also take in a CancellationToken now.

These changes involved a lot of change to how bytes are copied from 1 stream to another. The logic is more manual now in that it doesn't use the CopyToAsync method of the Stream class and instead copies in to a byte array (buffer) and then writes that data to the other stream. The current max buffer size (per iteration) is 80000 which was choosen because it is close to the default for Stream's "CopyToAsync" method of 81920.
Issue: https://github.com/TownSuite/TownSuite.IssueTracker/issues/7288