Fixed docker and zip file not working
Fixed the program not being able to be ran in Docker on a demo environment. The "/p:UseAppHost=false" seems to affect how the application uses certs for https.

Fixed zip file generation for when multiple png files need to be returned in a Stream. This was broken due seeking while ZipArchive is not disposed. ZipArchive was progressing the streams position between the seek and return of the Stream from the GetZipFileAsync method.
Issue: https://github.com/TownSuite/TownSuite.IssueTracker/issues/7288