Fixed warnings
Fixed all the warnings for the solution. Also refactored how the REPL makes a call to the API into more methods.

There were 4 warns. 3 were for possible nulls in the settings DTO for the repl, those values are now correctly validated to ensure they are not null before used. The launch settings for the APISite was using a string "true" instead of a boolean in the json.