Fixed pipeline build errors and warnings
Fixed an error when building in the pipeline due to no appsettings.Development.json file being found. That file is optional and intended to be different based on the development environment, so there is now a condition to ensure it exists before including it.

There were some warnings due to using "apt" instead of "apt-get". The warning is from code-factor and is as follows: "Do not use apt as it is meant to be a end-user tool, use apt-get or apt-cache instead (DL3027)".
Issue: https://github.com/TownSuite/TownSuite.IssueTracker/issues/7288